update: remove common components, move to customer components
