--- src/components/common/index.tsx
@@ -1,6 +1,0 @@
-import Logo from "./logo";
-import StarRating from "./starRating";
-
-export {
-  Logo, StarRating,
-};
