+++ src/components/customer/logo/index.tsx
