+++ src/components/customer/star-rating/index.tsx
